ok filters

+++ pizza.php
@@ -9,19 +9,26 @@
 
 	<?php 
 
-		if (isset($_GET['id']) && isset($_GET['validerId'])) {
+		if (isset($_GET['id']) && isset($_GET['validerId'])) { // ici pas besoin de filtrer je pense
 
-			$fichier = fopen($_GET['id'], 'r');
+			if (file_exists($_GET['id'])) {
 
-			$tab = JSON_decode(fgets($fichier));
+				$fichier = fopen($_GET['id'], 'r');
 
-			foreach($tab as $clef => $val)
-				echo $clef . " : " . $val . "<br>";
+				$tab = JSON_decode(fgets($fichier));
 
-			fclose($fichier);
+				foreach($tab as $clef => $val)
+					echo $clef . " : " . $val . "<br>";
 
-			echo "<a href='pizza.php'>Pages d'accueil</a>";
+				fclose($fichier);
 
+				echo "<a href='pizza.php'>Page d'accueil</a>";
+			}
+
+			else {
+				echo 'Identifiant de commande incorrect<br>'; 
+				echo "<a href='pizza.php'>Pages d'accueil</a>";
+			}
 		}
 
 		else if (isset($_POST['confirmer'])) {
@@ -37,25 +44,126 @@
 
 			echo "<b>Votre identifiant : " . $nom . "</b><br>";
 
-			echo "<a href='pizza.php'>Pages d'accueil</a>";
+			echo "<a href='pizza.php'>Page d'accueil</a>";
 			
 		}
 
-		else if (!isset($_POST['numTel']) 
-			  || !isset($_POST['mail']) 
-			  || !isset($_POST['adresse']) 
-			  || !isset($_POST['date']) 
-			  || !isset($_POST['heure']) 
-			  || !isset($_POST['pizza']) 
-			  || !isset($_POST['taille']) 
-			  || !isset($_POST['quantite']) 
-			  || !isset($_POST['sauce']) 
-			  || !isset($_POST['valider'])
-			  || $_POST['pizza'] == "Pizza Custom" && (!isset($_POST['ingredient']) || !isset($_POST['pate']))
-			  || $_POST['pizza'] == "Pizza Menu" && !isset($_POST['menu']) 
-			  || $_POST['pizza'] == "Pizza Menu" && (isset($_POST['ingredient']) || isset($_POST['pate']))
-			  || $_POST['pizza'] == "Pizza Custom" && isset($_POST['menu'])
-			  ) {
+		else if (isset($_POST['numTel']) 
+			  && isset($_POST['mail']) 
+			  && isset($_POST['adresse']) 
+			  && isset($_POST['date']) 
+			  && isset($_POST['heure']) 
+			  && isset($_POST['pizza']) 
+			  && isset($_POST['taille']) 
+			  && isset($_POST['quantite'])
+			  && isset($_POST['sauce']) 
+			  && isset($_POST['valider'])  // a partir de ces donditions on empeche les combinaisons incompatibles. Le mieux serait d'écrire un message d'erreur spécifique pour chacune d'entre elles décrivant l'erreur pour indiquer au client pourquoi ça commande n'est pas acceptée...
+			  && ($_POST['pizza'] == "Pizza Custom" && isset($_POST['ingredient']) && isset($_POST['pate'])
+			  || $_POST['pizza'] == "Pizza Menu" && isset($_POST['menu'])) 				// filtrage de pizza par la meme occasion
+			  && !($_POST['pizza'] == "Pizza Menu" && (isset($_POST['ingredient']) || isset($_POST['pate'])))
+			  && !($_POST['pizza'] == "Pizza Custom" && isset($_POST['menu']))) {
+
+			$_POST['adresse'] = filter_var($_POST['adresse'], FILTER_SANITIZE_STRING);	// on filtre adresse
+			$_POST['dare'] = filter_var($_POST['date'], FILTER_SANITIZE_STRING);		// on filtre date
+			$_POST['heure'] = filter_var($_POST['heure'], FILTER_SANITIZE_STRING);		// on filtre heure
+
+			if ($_POST['pizza'] == "Pizza Custom") {
+				foreach ($_POST['ingredient'] as $key => $value) 
+				$value = filter_var($value, FILTER_SANITIZE_STRING);					// on filtre ingrédiant si utilisateur modifie code HTML
+
+				$_POST['pate'] = filter_var($_POST['pate'], FILTER_SANITIZE_STRING);	// pareil
+			}
+			if ($_POST['pizza'] == "Pizza Menu")
+				$_POST['menu'] = filter_var($_POST['menu'], FILTER_SANITIZE_STRING);	// pareil
+
+			$_POST['numTel'] = filter_var($_POST['numTel'], FILTER_SANITIZE_STRING); 	 // numTel entre 10 et 15 chiffre
+			if (strlen($_POST['numTel']) < 10 || strlen($_POST['numTel']) > 15) {
+				echo "Numéro de téléphone incorrect.<br>";
+				echo "<a href='pizza.php'>Page d'accueil</a>";
+			}
+
+			else if (!filter_var($_POST['mail'], FILTER_VALIDATE_EMAIL)) { 				// on vérifie email valide
+				echo "Adresse email incorrecte<br>.";
+				echo "<a href='pizza.php'>Page d'accueil</a>";
+			}
+
+			else if (!($_POST['taille'] == "petite") && !($_POST['taille'] == "moyenne") && !($_POST['taille'] == "grande")) {
+				echo 'Taille incorrecte.<br>';											// on filtre taille
+				echo "<a href='pizza.php'>Page d'accueil</a>";
+			}
+
+			else if (!filter_var($_POST['quantite'], FILTER_VALIDATE_INT, array("options" => array("min_range" => 1, "max_range" => 30)))) {
+				echo "Quantite incorrecte.<br>";										// on filtre quantite entre 1 et 30
+				echo "<a href='pizza.php'>Page d'accueil</a>";
+			}
+
+			else {
+
+				echo "<b>Récapitulatif de la commande</b><br>";
+				echo $_POST['pizza'] . "<br>";
+				if ($_POST['pizza'] == "Pizza Menu")
+					echo $_POST['menu'];
+				else {
+					echo $_POST['pate'] . "<br>";
+					foreach ($_POST['ingredient'] as $val)
+						echo $val . " ";
+				}
+				echo "<br>" . $_POST['taille'] . "<br>";
+				echo "quantité : " . $_POST['quantite'] . "<br>";
+				echo "sauce ";
+				if ($_POST['sauce'] == "on")
+					echo "oui<br>";
+				else 
+					echo "non<br>";
+
+				$prix = 0;
+
+				if ($_POST['pizza'] == "Pizza Menu")
+					$prix += 10;
+				if ($_POST['pizza'] == "Pizza Custom")
+					$prix += 12;
+
+				if ($_POST['taille'] == "petite")
+					$prix /= (5/4);
+				if ($_POST['taille'] == "grande")
+					$prix *= (5/4);
+
+				$prix *= $_POST['quantite'];
+
+				echo "<b>Prix : " . $prix . " €</b>";
+
+				echo '
+					<form method="post" action="pizza.php">
+						<input type="hidden" name="numTel" value="' . $_POST['numTel'] . '">
+						<input type="hidden" name="mail" value="' . $_POST['mail'] .'">
+						<input type="hidden" name="adresse" value="' . $_POST['adresse'] .'">
+						<input type="hidden" name="date" value="' . $_POST['date'] .'">
+						<input type="hidden" name="heure" value="' . $_POST['heure'] .'">
+						<input type="hidden" name="pizza" value="' . $_POST['pizza'] .'">';
+						if ($_POST['pizza'] == "Pizza Menu")
+							echo '<input type="hidden" name="menu" value="' . $_POST['menu'] .'">';
+						if ($_POST['pizza'] == "Pizza Custom") {
+							echo '<input type="hidden" name="pate">';
+							$string = "";
+							foreach ($_POST['ingredient'] as $key => $value)
+								$string = $string . ' ' . $value;
+							echo  '<input type="hidden" name="ingredient" value="' . $string .'">';
+						}
+				echo '
+						<input type="hidden" name="taille" value="' . $_POST['taille'] .'">
+						<input type="hidden" name="quantite" value="' . $_POST['quantite'] .'">
+						<input type="hidden" name="sauce" value="' . $_POST['sauce'] .'">
+						<input type="hidden" name="prix" value="' . $prix . '">
+
+						<input type="submit" name="confirmer" value="confirmer">
+						<a href="pizza.php">Annuler</a>
+					</form>
+				';
+			}
+			
+		}
+
+		else {
 
 			echo '
 
@@ -67,7 +175,6 @@
 					</fieldset><br>
 				</form>';
 
-			echo "<b>Veuillez remplir votre commande correctement s'il vous plait</b><br>";
 			echo '
 
 				<form method="post" action="pizza.php">
@@ -183,69 +290,6 @@
 					<input value="Remettre à zéro" type="reset" name="zero">
 				</form>
 
-			';
-		}
-
-		else {
-			echo "<b>Récapitulatif de la commande</b><br>";
-			echo $_POST['pizza'] . "<br>";
-			if ($_POST['pizza'] == "Pizza Menu")
-				echo $_POST['menu'];
-			else {
-				echo $_POST['pate'] . "<br>";
-				foreach ($_POST['ingredient'] as $val)
-					echo $val . " ";
-			}
-			echo "<br>" . $_POST['taille'] . "<br>";
-			echo "quantité : " . $_POST['quantite'] . "<br>";
-			echo "sauce ";
-			if ($_POST['sauce'] == "on")
-				echo "oui<br>";
-			else 
-				echo "non<br>";
-
-			$prix = 0;
-
-			if ($_POST['pizza'] == "Pizza Menu")
-				$prix += 10;
-			if ($_POST['pizza'] == "Pizza Custom")
-				$prix += 12;
-
-			if ($_POST['taille'] == "petite")
-				$prix /= (5/4);
-			if ($_POST['taille'] == "grande")
-				$prix *= (5/4);
-
-			$prix *= $_POST['quantite'];
-
-			echo "<b>Prix : " . $prix . " €</b>";
-
-			echo '
-				<form method="post" action="pizza.php">
-					<input type="hidden" name="numTel" value="' . $_POST['numTel'] . '">
-					<input type="hidden" name="mail" value="' . $_POST['mail'] .'">
-					<input type="hidden" name="adresse" value="' . $_POST['adresse'] .'">
-					<input type="hidden" name="date" value="' . $_POST['date'] .'">
-					<input type="hidden" name="heure" value="' . $_POST['heure'] .'">
-					<input type="hidden" name="pizza" value="' . $_POST['pizza'] .'">';
-					if ($_POST['pizza'] == "Pizza Menu")
-						echo '<input type="hidden" name="menu" value="' . $_POST['menu'] .'">';
-					if ($_POST['pizza'] == "Pizza Custom") {
-						echo '<input type="hidden" name="pate">';
-						$string = "";
-						foreach ($_POST['ingredient'] as $key => $value)
-							$string = $string . ' ' . $value;
-						echo  '<input type="hidden" name="ingredient" value="' . $string .'">';
-					}
-			echo '
-					<input type="hidden" name="taille" value="' . $_POST['taille'] .'">
-					<input type="hidden" name="quantite" value="' . $_POST['quantite'] .'">
-					<input type="hidden" name="sauce" value="' . $_POST['sauce'] .'">
-					<input type="hidden" name="prix" value="' . $prix . '">
-
-					<input type="submit" name="confirmer" value="confirmer">
-					<a href="pizza.php">Annuler</a>
-				</form>
 			';
 		}
